parte1/requirements.txt
alembic==1.16.5
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
duckdb==1.3.2
duckdb_engine==0.17.0
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.59.2
greenlet==3.2.4
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupysql==0.11.1
jupysql-plugin==0.4.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
narwhals==2.5.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.3
optuna==4.5.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
ploomber-core==0.2.27
plotly==5.15.0
polars==1.33.1
posthog==6.7.4
prettytable==3.16.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.43
sqlglot==27.14.0
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
